% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{boot}
\alias{boot}
\title{Multinomial state-space bootstrap}
\usage{
boot(object, ...)
}
\arguments{
\item{object}{An object for which the bootstrap method is defined.}

\item{...}{Additional arguments passed to the method.}
}
\description{
This is a generic bootstrap function for models.
}
