% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{coef.mnTS}
\alias{coef.mnTS}
\title{Multinomial state-space coefficient summary}
\usage{
\method{coef}{mnTS}(mod, ...)
}
\arguments{
\item{mod}{inout matrix of}

\item{...}{input matrix of}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function
customize the text to print
}
