% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnGLMM.R
\name{mnGLMM.ml}
\alias{mnGLMM.ml}
\title{Multinomial GLMM Maximum likelihood}
\usage{
mnGLMM.ml(par, par.fixed, Y, X, fitted.values = FALSE, REML = FALSE)
}
\description{
This is an internal function called by mnGLMM.
}
