% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{mnTS.ml}
\alias{mnTS.ml}
\title{Multinomial state-space Maximum likelihood}
\usage{
mnTS.ml(par, par.fixed, Y, X = NULL, Tsample, fitted.values = FALSE)
}
\description{
This is an internal function called by mnTS
}
