% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS_wrapper.R
\name{mnTS.ml.wrapper}
\alias{mnTS.ml.wrapper}
\title{mnTS Wrapper}
\usage{
mnTS.ml.wrapper(par, par.fixed, Y, X = NULL, Tsample)
}
\arguments{
\item{par}{something}

\item{par.fixed}{More things}

\item{Y}{inout matrix of}

\item{X}{input matrix of}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function that, by default, prints "Hello world". You can
customize the text to print (using the \code{to_print} argument) and add
an exclamation point (\code{excited = TRUE}).
}
\examples{
hello_world()
hello_world(excited = TRUE)
hello_world(to_print = "Hi world")

}
