% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS_wrapper.R
\name{mnTS_ml_cpp_listout}
\alias{mnTS_ml_cpp_listout}
\title{mnTS_ml_cpp_list}
\usage{
mnTS_ml_cpp_listout(par, par_fixed, Y, X, Tsample)
}
\description{
This is a description of what the function does.
}
