% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{story_100_bootstraps}
\alias{story_100_bootstraps}
\title{Story 100 bootstraps}
\format{
A list containing the bootstrapped outputs.
}
\usage{
story_100_bootstraps
}
\description{
100 bootstraps of the story data as modelled in the vignette
}
\examples{
data(story_100_bootstraps)
}
\keyword{datasets}
