% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{story_char_matrix}
\alias{story_char_matrix}
\title{Story Charcoal Matrix}
\format{
A `1` column matrix with `160` rows (observations).
}
\source{
For details see vignette
}
\usage{
story_char_matrix
}
\description{
A matrix of charcoal accumulation rates from Story Lake, Indiana.
   This matrix has been binned at a 50-year resolution and scaled.
}
\examples{
data(story_char_matrix)
head(story_char_matrix)
}
\keyword{datasets}
