% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{story_char_wide}
\alias{story_char_wide}
\title{Story Charcoal Wide}
\format{
A tibble with `736` rows and `3` columns.
}
\source{
For details see https://doi.org/10.1111/1365-2745.14289
}
\usage{
story_char_wide
}
\description{
A ~10,000-year data set from Story Lake, Indiana, of modelled ages, sample
   depths, and charcoal accumulation rate from a sediment core sample.
}
\examples{
data(story_char_wide)
str(story_char_wide)
}
\keyword{datasets}
