% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnGLMM.R
\name{summary.mnGLMM}
\alias{summary.mnGLMM}
\title{GLMM summary}
\usage{
\method{summary}{mnGLMM}(mod, ...)
}
\arguments{
\item{mod}{An object of class "mnGLMM"}
}
\value{
Prints coefficient estimates from a fitted mnGLMM
}
\description{
Prints the summary of a mnGLMM model.
}
