% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{summary.mnTS}
\alias{summary.mnTS}
\title{Summary Method for Multinomial State-Space Models}
\usage{
\method{summary}{mnTS}(mod, ...)
}
\arguments{
\item{mod}{An object of class \code{"mnTS"}, typically returned by the
\code{\link{mnTS}} function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Returns a summary list containing key model estimates.
    Also prints the summary to the console.
}
\description{
Provides a summary of the fitted multinomial state-space model.
}
