% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{boot.mnTS}
\alias{boot.mnTS}
\title{Multinomial state-space bootstrap}
\usage{
boot.mnTS(mod, reps, dispersion.fixed = 1, maxit.optim = 1e+05, ...)
}
\arguments{
\item{mod}{inout matrix of}

\item{...}{input matrix of}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function
}
