% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{coef.mnTS}
\alias{coef.mnTS}
\title{Coefficients for Multinomial State-Space Model}
\usage{
\method{coef}{mnTS}(mod, ...)
}
\arguments{
\item{mod}{An object of class \code{"mnTS"}, as returned by the
\code{\link{mnTS}} function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of estimated model coefficients.
}
\description{
Extracts estimated coefficients from a fitted multinomial state-space model.
}
