% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnGLMM.R
\name{mnGLMM.ml}
\alias{mnGLMM.ml}
\title{Multinomial GLMM Maximum likelihood}
\usage{
mnGLMM.ml(par, par.fixed, Y, X, fitted.values = FALSE, REML = FALSE)
}
\arguments{
\item{par}{something}

\item{par.fixed}{More things}

\item{Y}{inout matrix of}

\item{X}{input matrix of}

\item{fitted.values}{TRUE/FALSE argument indicating}

\item{REML}{Description}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function that, by default, prints "Hello world". You can
customize the text to print (using the \code{to_print} argument) and add
an exclamation point (\code{excited = TRUE}).
}
\examples{
hello_world()
hello_world(excited = TRUE)
hello_world(to_print = "Hi world")

}
