% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnGLMM.R
\name{mnGLMM}
\alias{mnGLMM}
\title{Multinomial Generalized Linear Mixed Model (mnGLMM)}
\usage{
mnGLMM(
  Y,
  X = NULL,
  B.fixed = if (is.null(X)) matrix(c(0, rep(NA, ncol(Y) - 1)), nrow = 1, ncol = ncol(Y))
    else matrix(c(rep(0, (ncol(X) + 1) * ncol(Y)), rep(NA, (ncol(Y) - 1) * (ncol(X) +
    1))), nrow = ncol(X) + 1, ncol = ncol(Y)),
  B.start = if (is.null(X)) matrix(0, nrow = 1, ncol = ncol(Y)) else matrix(0, nrow =
    ncol(X) + 1, ncol = ncol(Y)),
  sigma.fixed = NA,
  sigma.start = 0.1,
  dispersion.fixed = 1,
  dispersion.start = 1,
  V.fixed = diag(ncol(Y)),
  V.start = diag(ncol(Y)),
  method = "bobyqa",
  optim.control = NULL,
  maxit.optim = 1e+05,
  REML = FALSE,
  compute.information.matrix = TRUE
)
}
\arguments{
\item{Y}{A matrix of multinomially distributed count data
(e.g., community count data).}

\item{X}{A matrix of covariates (predictors), which may be of mixed type
Covariates should be scaled when appropriate. Can be \code{NULL}.}

\item{B.fixed}{A matrix indicating which B coefficients
(driver-species relationships) to estimate. The number of columns must
equal \code{ncol(Y)} (number of species), and the number of rows must
equal \code{ncol(X) + 1} (intercept + covariates).}

\item{B.start}{A matrix of starting values for the B coefficients.
Dimensions must match \code{B.fixed}.}

\item{sigma.fixed}{Fixed value for the overall model variance. Use \code{NA}
to estimate it from the model.}

\item{sigma.start}{Starting value for estimating \code{sigma.fixed}
(default is 0.1).}

\item{dispersion.fixed}{Fixed dispersion parameter to account for
over- or under-dispersion. A value of 1 corresponds to no
extra dispersion (pure multinomial).}

\item{dispersion.start}{Starting value for estimating
\code{dispersion.fixed}.}

\item{V.fixed}{A species-by-species covariance matrix representing
environmental variation.}

\item{V.start}{Starting values for \code{V.fixed}.}

\item{method}{Optimization method. Acceptable values include
\code{"Nelder-Mead"}, \code{"BFGS"} (via \code{optim}),
and \code{"bobyqa"} (via the \code{minqa} package).}

\item{optim.control}{Optional list of control parameters passed to the
optimizer. See the \code{minqa} package documentation for details.}

\item{maxit.optim}{Maximum number of iterations for the optimizer
(default is 1e+05). Increase if of optimizer needs more itterations.}

\item{REML}{Logical. If \code{TRUE}, uses restricted maximum likelihood
for variance estimation.}

\item{compute.information.matrix}{Logical. If \code{TRUE},
computes the observed information matrix.}
}
\value{
An object of class \code{"mnGLMM"}.
}
\description{
Fits a multivariate generalized linear regression model for multinomially
distributed response data. This function estimates driver-species
relationships (B coefficients), observation-level dispersion,
and species covariance structures via maximum likelihood or REML.
}
