% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{mnTS}
\alias{mnTS}
\title{Multinomial state-space}
\usage{
mnTS(
  Y,
  X = NULL,
  Tsample = 1:nrow(Y),
  B0.fixed = matrix(c(0, rep(NA, ncol(Y) - 1)), nrow = 1, ncol = ncol(Y)),
  B0.start = matrix(0, nrow = 1, ncol = ncol(Y)),
  C.fixed = diag(rep(NA, ncol(Y))),
  C.start = 0.01 * diag(ncol(Y)),
  B.fixed = if (is.null(X)) NULL else matrix(NA, nrow = ncol(X), ncol = ncol(Y)),
  B.start = if (is.null(X)) NULL else matrix(0, nrow = ncol(X), ncol = ncol(Y)),
  sigma.fixed = NA,
  sigma.start = 0.1,
  dispersion.fixed = 1,
  dispersion.start = 1,
  V.fixed = diag(ncol(Y)),
  V.start = diag(ncol(Y)),
  compute.information.matrix = TRUE,
  method = "bobyqa",
  optim.control = NULL,
  maxit.optim = 1e+05,
  hessian.method.args = list(eps = 1e-04, d = 1e-04, r = 4, v = 2)
)
}
\arguments{
\item{Y}{inout matrix of}

\item{X}{input matrix of}

\item{Tsample}{Placement of observations}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function that, by default, prints "Hello world". You can
customize the text to print (using the \code{to_print} argument) and add
an exclamation point (\code{excited = TRUE}).
}
\examples{
hello_world()
hello_world(excited = TRUE)
hello_world(to_print = "Hi world")

}
