% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{simulate.mnTS}
\alias{simulate.mnTS}
\title{Simulate Data from a Multinomial State-Space Model}
\usage{
\method{simulate}{mnTS}(mod, ...)
}
\arguments{
\item{mod}{An object of class \code{"mnTS"}, as returned by the
\code{\link{mnTS}} function.}

\item{...}{Additional arguments passed to the simulation routine.}
}
\value{
A simulated dataset in the same structure as the response variable
    \code{Y} used to fit the model.
}
\description{
Generates simulated datasets from a fitted mnTS object.
}
