% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnGLMM.R
\name{summary.mnGLMM}
\alias{summary.mnGLMM}
\title{GLMM summary}
\usage{
\method{summary}{mnGLMM}(mod, ...)
}
\arguments{
\item{mod}{inout matrix of}

\item{...}{input matrix of}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function
customize the text to
}
