% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_mnTS.R
\name{summary.mnTS}
\alias{summary.mnTS}
\title{Multinomial state-space summary}
\usage{
\method{summary}{mnTS}(mod, ...)
}
\arguments{
\item{mod}{inout matrix of}
}
\value{
This function returns a phrase to print, with or without an
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function
customize the text to print (using the \code{to_print} argument) and add
an exclamation point (\code{excited = TRUE}).
}
